+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Security
+- upgrade drupal/core 10.0.11 => 10.2.5 - QS-67
+
+### Added
+- add module jquery_ui_draggable - QS-67
 
 ## [3.0.0] - 2024-04-22
 ### Security