+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Security
+- upgrade drupal/core 10.0.11 => 10.2.5 - QS-67
+
+### Added
+- add module jquery_ui_draggable - QS-67
+- mark all custom modules & themes as Drupal 11 compliant - QS-67
+
+### Fixed
+- remove usage of deprecated ContainerAware class - QS-67
+
+### Changed
+- replace usage of deprecated trait Drupal\Tests\field\Traits\EntityReferenceTestTrait in favor of EntityReferenceFieldCreationTrait - QS-67
 
 ## [3.0.0] - 2024-04-22
 ### Security