+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.1] - 2024-05-06
+### Security
+- upgrade drupal/core 10.0.11 => 10.2.5 - QS-67
+
+### Added
+- add module jquery_ui_draggable - QS-67
+- mark all custom modules & themes as Drupal 11 compliant - QS-67
+
+### Fixed
+- remove usage of deprecated ContainerAware class - QS-67
+
+### Changed
+- replace usage of deprecated trait Drupal\Tests\field\Traits\EntityReferenceTestTrait in favor of EntityReferenceFieldCreationTrait - QS-67
+
 ## [3.0.0] - 2024-04-22
 ### Security
 - upgrade drupal/core 9.5.11 => 10.0.11 - QS-50
@@ -717,7 +731,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - init empty repo
 
-[Unreleased]: https://github.com/antistatique/quartiers-solidaires/compare/3.0.0...HEAD
+[Unreleased]: https://github.com/antistatique/quartiers-solidaires/compare/3.0.1...HEAD
+[3.0.1]: https://github.com/antistatique/quartiers-solidaires/compare/3.0.0...3.0.1
 [3.0.0]: https://github.com/antistatique/quartiers-solidaires/compare/2.3.8...3.0.0
 [2.3.8]: https://github.com/antistatique/quartiers-solidaires/compare/2.3.7...2.3.8
 [2.3.7]: https://github.com/antistatique/quartiers-solidaires/compare/2.3.6...2.3.7